Homework 4

+++ src/main/java/org/kaoden/ws/homework/service/entry/EntryServiceImpl.java
@@ -1,17 +1,19 @@
-package org.kaoden.ws.homework.service;
+package org.kaoden.ws.homework.service.entry;
 
 import lombok.RequiredArgsConstructor;
+import org.kaoden.ws.homework.exception.NotFoundException;
 import org.kaoden.ws.homework.model.Entry;
-import org.kaoden.ws.homework.repository.EntryRepository;
-import org.kaoden.ws.homework.service.argument.CreateEntryArgument;
-import org.kaoden.ws.homework.service.argument.UpdateEntryArgument;
+import org.kaoden.ws.homework.repository.entry.EntryRepository;
+import org.kaoden.ws.homework.service.entry.argument.CreateEntryArgument;
+import org.kaoden.ws.homework.service.entry.argument.UpdateEntryArgument;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class EntryServiceImp implements EntryService {
+public class EntryServiceImpl implements EntryService {
 
     private final EntryRepository repository;
 
@@ -27,7 +29,8 @@ public class EntryServiceImp implements EntryService {
 
     @Override
     public Entry getExisting(Long id) {
-        return repository.findById(id);
+        return Optional.ofNullable(repository.findById(id))
+                       .orElseThrow(() -> new NotFoundException("There is no entry with this ID: " + id));
     }
 
     @Override
@@ -37,16 +40,28 @@ public class EntryServiceImp implements EntryService {
 
     @Override
     public Entry update(Long id, UpdateEntryArgument entry) {
+        if (! exists(id)) {
+            throw new NotFoundException("Impossible update entry with this ID: " + id);
+        }
         return repository.update(id, Entry.builder()
                                           .id(id)
                                           .name(entry.getName())
                                           .description(entry.getDescription())
                                           .link(entry.getLink())
                                           .build());
+
     }
 
     @Override
     public void delete(Long id) {
+        if (! exists(id)) {
+            throw new NotFoundException("Impossible delete entry with this ID: " + id);
+        }
         repository.delete(id);
+    }
+
+    @Override
+    public Boolean exists(Long id) {
+        return repository.exists(id);
     }
 }
